+++ src/tokenizing.c
@@ -22,36 +22,17 @@ int	ft_quotecho_condition(char **val, int y, char *str, int *i)
 	return (y);
 }
 
-int	ft_tok_type(char *value)
-{
-	int	type;
-
-	type = 0;
-	if (value[0] == ' ')
-		type = SEP;
-	else if (value[0] == '>')
-		type = REDIR_OUT;
-	else if (value[0] == '<')
-	{
-		if (value[1] && value[1] == '<')
-			type = HERE_DOC;
-		else
-			type = REDIR_IN;
-	}
-	else
-		type = COM;
-	return (type);
-}
-
-//TODO: too many arguements here <3
-t_tokens	*ft_new_echo_token(int i, int ind, int *y, char **strs, int redir)
+t_tokens	*ft_new_token(int i, int *y, char **strs, int redir)
 {
 	t_tokens	*new_tok;
 
 	new_tok = malloc(sizeof(t_tokens));
 	if (!new_tok)
 		return (NULL);
-	new_tok->val = ft_tok_val(strs[i], y, 1, redir);
+	if (redir == -1)
+		new_tok->val = ft_tok_val(strs[i], y, 0, 0);
+	else
+		new_tok->val = ft_tok_val(strs[i], y, 1, redir);
 	if (!new_tok->val)
 	{
 		free(new_tok);
@@ -59,11 +40,20 @@ t_tokens	*ft_new_echo_token(int i, int ind, int *y, char **strs, int redir)
 	}
 	new_tok->type = ft_tok_type(new_tok->val);
 	new_tok->ind_command = i;
-	new_tok->ind_word = ind;
 	return (new_tok);
 }
 
-//TODO: overline
+void	ft_if_echo_redir(int *echo, int *redir, t_tokens *toks, int ind)
+{
+	if (toks->type == COM && !ft_strncmp(toks->val, "echo", 4)
+		&& ind < 2 && ft_strlen(toks->val) == 4)
+		(*echo)++;
+	if (toks->type == REDIR_OUT || toks->type == REDIR_IN)
+		*redir = 1;
+	if (*redir && toks->type == COM)
+		*redir = 0;
+}
+
 int	ft_loop_new_com_tok(char **strs, t_tokens *toks, int i, int y)
 {
 	int	ind;
@@ -77,17 +67,11 @@ int	ft_loop_new_com_tok(char **strs, t_tokens *toks, int i, int y)
 	redir = 0;
 	while (strs[i][y])
 	{
-		if (toks->type == COM && !ft_strncmp(toks->val, "echo", 4)
-			&& ind < 2 && ft_strlen(toks->val) == 4)
-			echo++;
-		if (toks->type == REDIR_OUT || toks->type == REDIR_IN)
-			redir = 1;
-		if (redir && toks->type == COM)
-			redir = 0;
+		ft_if_echo_redir(&echo, &redir, toks, ind);
 		if (echo)
-			toks->next = ft_new_echo_token(i, ind, &y, strs, redir);
+			toks->next = ft_new_token(i, &y, strs, redir);
 		else
-			toks->next = ft_new_token(i, ind, &y, strs);
+			toks->next = ft_new_token(i, &y, strs, -1);
 		if (!toks->next)
 			return (1);
 		ft_assign_prev_cur_tok(&toks);
@@ -104,7 +88,7 @@ t_tokens	*ft_gimme_tokens(char **strs)
 	t_tokens	*head;
 
 	y = 0;
-	toks = ft_new_token(0, 0, &y, strs);
+	toks = ft_new_token(0, &y, strs, -1);
 	if (!toks)
 		return (ft_free_toks(toks));
 	toks->prev = NULL;
